requirements.txt
numpy == 1.24.2
pandas == 1.5.3
scikit-learn == 1.2.1
nltk == 3.8.1
fastapi == 0.92.0
uvicorn == 0.20.0
requests == 2.28.2
# for aws lambda
mangum == 0.17.0

# for ml
scikit-surprise == 1.1.3